feat(layout): move layout and tasklist sidecars to own files
