+++ backend/tests/services/test_task_service.py
@@ -10,7 +10,8 @@ from backend.app.models.project import Project
 from backend.app.models.task import TaskPriority, TaskStatus
 from backend.app.models.validation import ValidationError, ValidationResult
 from backend.app.services.task_service import TaskService
-from src.model import ShapeShiftProject, TaskList
+from src.model import ShapeShiftProject
+from src.task_list import TaskList
 
 # pylint: disable=redefined-outer-name, unused-argument
